refactor format money function and added settings feature

+++ src/features/budgets/budget-manage/BudgetManage.tsx
@@ -54,7 +54,7 @@ const BudgetManage = () => {
 
       const fields = {
         ...formData,
-        amount: parseInt(formData.amount),
+        amount: parseFloat(formData.amount),
       };
 
       if (selectedBudget) {
